requirements.txt
PyGithub
PyYAML
requests
pyOpenSSL
